Full comps table with prices and as reported fundamentals data from IEX Cloud
Full description included on the Instructions tab.  This spreadsheet uses Webservice and PowerQuery - and will work with a free IEX CLoud account, and works even better with a paid account.  I make some adjustments to the as reported xbrl items (as submitted to the SEC) so that companies are somewhat more comparable.  These are documents in the Adjustments Tab.  Includes a traditional comps view, as well as "sidebyside" views of the data.